+++ src/main/java/DepositProcessor.java
@@ -7,6 +7,6 @@ public class DepositProcessor {
 
     public void process(String command) {
         String[] words = command.split("\\s+");
-        this.bank.depositIntoAccount(Integer.parseInt(words[1]), Float.parseFloat(words[2]));
+        this.bank.depositIntoAccount(Integer.parseInt(words[1]), Double.parseDouble(words[2]));
     }
 }
